scripts: split LLBC extraction into engine + drivers
Move the extraction machinery into an import-only engine,
scripts/llbc_extract.py (CrateSpec + run_cli), carrying no crate
names. Add pyre/scripts/extract-llbc.py declaring the pyre crate
table (corpus, pyre-object, pyre-module, pyre-interpreter, pyre-jit;
default trio pyre-object/pyre-interpreter/pyre-jit). scripts/extract-llbc.py
becomes a forwarding shim to the pyre driver, kept because CI
workflows, pyre/pyre-jit-trace/build.rs, and pyre/check.py invoke it
by that path.

Remove the aheui-specific census test from majit-translate and
reword two comments (lib.rs, local_crates.rs) that named the external
consumer crate to generic wording.
